re-format vector implementation
The var-vector and fix-vector are re-implemented. The implementation for
double-value matrix is included in vector classes. When some functions
which are not suitable for non-double types are invoked for non-double
type, compiling errors would occur. These errors are not implemented
with assert as before. This kinf of implementation simplifies codes.