add interaction for set of lists
add a wrapper for set of lists (simply var-vector, using a class rather
than typedef since there are templates). This works for interaction
calculation for various kinds of interactions. Here we assumed that the
common distance matrix and geometry are used (though in principle they
could be different for various interactions).