Finish the propagator framework
In this commit, we did two things, one is related to type check for
vector type. The string-type type tag is removed and int-type flag is
added into vector-type class. the check if the type is vector-type is
implemented by checking if the class has this variable. The other is
finished the test for propagator.