finish the dsfmt generator
The dsfmt random generator is finished. The test is also passed.
However, SSE2-related codes may produce errors due to alignment
when using pipeline to output (such as xxx > 1). It could be ok
when using the form xxx > 1 2> 2. The reason needs more investigations.
The no-SSE2 version is OK. a comparison with previous codes may be
needed.