reform the propagator
the parameter structure are reformed. init and clear of parameter
are introduced. However, some data related to global data format cannot
be allocated in init operation. and those data should be allocated and
released manually, better just introduce a pointer rather than to
allocate the pointer in unique parameter directly.