Add Definition PDB for Residue Key
Since the key depends on the sequence of atom name, addition of new
atoms would change the key. To validate the produce of key library, a
set of definitions of residues (in not-precise PDB format) is used to
stored supported residues and their atomic definitions. This is not
finished. A problem occurs for MET in 2RNM (the Atom H is missing).