update PDB-File object
PDB-File operation is updated, especially for AltLoc flag. Since the
AltLoc definition is not very clear (such as if all Altloc are
correlated? for some altloc flag, how to choose the default
conformations?) I implement for regular cases. there are definitely
exceptions which would not be considered in this code.