add complete expression-sum
all possible expressions for sum are added. Due to the complexity of
input object and the complexity of specialization, three kinds of input
objects are used (constant, array, and two-member array-operation). This
could be applied to other expressions.

besides, some other expressions and array operations are added, but not
complete.