add detailed interfaces for short-data array
many detailed operations for short-data array are added to matching
various conditions for efficiency. (The complexity of coding may be
originated from the problem of template definition.) maybe I need to
know more skills for C++ GP. maybe code generator is a kind of solution.
(however, it is out of the language level).

more implementations for operator /= of short-data array are needed.
more implementations for large-data array are also needed.